# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API for managing artists, albums, and songs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Retrieve all albums",
                "description": "Returns a list of all albums",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "responses": {
                    "200": {
                        "description": "List of albums",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Album"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create a new albums",
                "description": "Creates a new album and returns the created album object",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "release_year",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Name of the Album",
                                        "type": "string",
                                        "example": "Best Ever"
                                    },
                                    "release_year": {
                                        "description": "Year of album release",
                                        "type": "integer",
                                        "example": 1992
                                    },
                                    "artist_id": {
                                        "description": "Identifier of artist that album belongs to",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Retrieve an album by ID",
                "description": "Fetches an album by their unique identifier",
                "operationId": "71e9c4179582355e89419e6f6994c418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the album to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Update an album",
                "description": "Updates an existing album by ID",
                "operationId": "1300ec85cc590b1ae1468c9bf8db122d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the album to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "release_year",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Name of the Album",
                                        "type": "string",
                                        "example": "Best Ever"
                                    },
                                    "release_year": {
                                        "description": "Year of album release",
                                        "type": "integer",
                                        "example": 1992
                                    },
                                    "artist_id": {
                                        "description": "Identifier of artist that album belongs to",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Delete an album",
                "description": "Deletes a album by ID",
                "operationId": "cabfc7c152115607d310827d46a82d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the album to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Album successfully deleted"
                    },
                    "404": {
                        "description": "Album not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Album": {
                "title": "Album",
                "description": "Schema for Album",
                "required": [
                    "id",
                    "title",
                    "release_year",
                    "artist_id"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the song",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Name of the Album",
                        "type": "string",
                        "example": "Best Ever"
                    },
                    "release_year": {
                        "description": "Year of album release",
                        "type": "integer",
                        "example": 1992
                    },
                    "artist_id": {
                        "description": "Identifier of artist that album belongs to",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date when the album was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date when the album was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}